Extracted helper functions from average to utility to prepare similar concat usage.
